+++ Tests/TestRunners/SelfVerifyingTestFactory.cs
@@ -42,7 +42,7 @@ namespace CodeConverter.Tests.TestRunners
         {
             // Lazy to avoid confusing test runner on error, but also avoid calculating multiple times
             var conversionResultAsync = new AsyncLazy<ConversionResult>(() =>
-                ProjectConversion.ConvertText<TLanguageConversion>(sourceFileText, DefaultReferences.NetStandard2)
+                ProjectConversion.ConvertText<TLanguageConversion>(sourceFileText, new TextConversionOptions(DefaultReferences.NetStandard2))
             );
 
             var runnableTestsInTarget = new AsyncLazy<Dictionary<string, NamedFact>>(async () => GetConvertedNamedFacts<TTargetCompiler>(runnableTestsInSource,